requirements.txt
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
ipython==8.22.1